More sample data
Added order_products, refactored seed data into separate file.
